backend/requirements.txt
awscli==1.38.27
boto3==1.37.27
botocore==1.37.27
colorama==0.4.6
dnspython==2.7.0
docutils==0.16
jmespath==1.0.1
pillow==11.1.0
pyasn1==0.6.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
